chore(refactor): Refactor of Data Collection modules to match ETL workflow
